pyproject.toml
[project]
name = "hamsfly-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.1.7",
    "chromadb>=1.1.1",
    "google-genai>=1.45.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
